+++ WebApp/WebApp/Controllers/PersonController.cs
@@ -3,6 +3,8 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using WebApp.Models;
 using WebApp.Repositories;
+using WebApp.Services;
+using WebApp.Validation;
 
 namespace WebApp.Controllers
 {
@@ -12,10 +14,15 @@ namespace WebApp.Controllers
     public class PersonController : ControllerBase
     {
         public readonly PersonRepository personRepository;
-       
-        public PersonController(PersonRepository personRepository)
+        private IPersonValidator personValidator;
+        private IPersDelValidator personDelValidator;
+        public PersonController(PersonRepository personRepository,
+            IPersonValidator personValidator,
+             IPersDelValidator personDelValidator)
         {
             this.personRepository = personRepository;
+            this.personValidator = personValidator;
+            this.personDelValidator = personDelValidator;
         }
 
         [HttpGet]
@@ -29,8 +36,13 @@ namespace WebApp.Controllers
         public IActionResult Create([FromBody] PersonModel newPerson)
 
         {
+            var validation = new OperationResult<PersonModel>(personValidator.ValidateEntity(newPerson));
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+            }
             personRepository.Add(newPerson);
-            return Ok(newPerson);
+            return NoContent();
         }
 
         [HttpPut]
@@ -44,6 +56,12 @@ namespace WebApp.Controllers
         [HttpDelete]
         public IActionResult Delete([FromBody] int Id)
         {
+            var request = new PersonDeleteModel { Id = Id };
+            var validation = new OperationResult<PersonDeleteModel>(personDelValidator.ValidateEntity(request));
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+            }
             personRepository.Delete(Id);
             return NoContent();
         }
